requirements.txt
asgiref==3.2.7
autopep8==1.5.2
Django==3.0.6
django-extensions==2.2.9
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.1
pycodestyle==2.5.0
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
uuid==1.30
